api/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
psycopg2-binary==2.9.1
django-cors-headers==3.8.0
django-environ==0.4.5
